bevobeacon-iaq/requirements.txt
pandas
numpy 

# Python Sensor Libraries
adafruit-circuitpython-tsl2591
adafruit-circuitpython-tsl2561
adafruit-circuitpython-sgp30
adafruit-circuitpython-pcf8523
sps30
crcmod
pyserial

# AWS Libraries
# boto3

#I2C tools
smbus2
smbus
scd30